Remove AE2 asm and use the new API instead
