--- src/main/java/com/github/vfyjxf/nee/asm/NEELoadingPlugin.java
@@ -1,33 +1,0 @@
-package com.github.vfyjxf.nee.asm;
-
-import java.util.Map;
-
-import cpw.mods.fml.relauncher.IFMLLoadingPlugin;
-
-@IFMLLoadingPlugin.Name("NEELoadingPlugin")
-@IFMLLoadingPlugin.MCVersion("1.7.10")
-public class NEELoadingPlugin implements IFMLLoadingPlugin {
-
-    @Override
-    public String[] getASMTransformerClass() {
-        return new String[] { "com.github.vfyjxf.nee.asm.NEEClassTransformer" };
-    }
-
-    @Override
-    public String getModContainerClass() {
-        return null;
-    }
-
-    @Override
-    public String getSetupClass() {
-        return null;
-    }
-
-    @Override
-    public void injectData(Map<String, Object> data) {}
-
-    @Override
-    public String getAccessTransformerClass() {
-        return null;
-    }
-}
