+++ src/main/java/com/github/vfyjxf/nee/client/GuiEventHandler.java
@@ -2,8 +2,11 @@ package com.github.vfyjxf.nee.client;
 
 import static com.github.vfyjxf.nee.config.NEEConfig.draggedStackDefaultSize;
 import static com.github.vfyjxf.nee.config.NEEConfig.useStackSizeFromNEI;
+import static com.github.vfyjxf.nee.nei.NEECraftingHandler.INPUT_KEY;
 import static com.github.vfyjxf.nee.nei.NEECraftingHelper.tracker;
+import static com.github.vfyjxf.nee.utils.GuiUtils.isPatternTerm;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
@@ -17,26 +20,33 @@ import net.minecraft.item.ItemStack;
 import net.minecraftforge.client.event.GuiOpenEvent;
 import net.minecraftforge.client.event.GuiScreenEvent;
 
+import org.lwjgl.input.Keyboard;
 import org.lwjgl.input.Mouse;
-
-import appeng.client.gui.AEBaseGui;
-import appeng.client.gui.implementations.GuiCraftConfirm;
-import appeng.client.gui.implementations.GuiPatternTerm;
-import appeng.container.implementations.ContainerCraftConfirm;
-import appeng.container.implementations.ContainerPatternTerm;
-import appeng.container.slot.SlotFake;
-import codechicken.nei.VisiblityData;
-import codechicken.nei.api.INEIGuiHandler;
-import codechicken.nei.api.TaggedInventoryArea;
 
 import com.github.vfyjxf.nee.client.gui.widgets.GuiImgButtonEnableCombination;
 import com.github.vfyjxf.nee.config.ItemCombination;
 import com.github.vfyjxf.nee.config.NEEConfig;
 import com.github.vfyjxf.nee.container.ContainerCraftingConfirm;
+import com.github.vfyjxf.nee.nei.NEECraftingHandler;
 import com.github.vfyjxf.nee.network.NEENetworkHandler;
 import com.github.vfyjxf.nee.network.packet.PacketSlotStackChange;
+import com.github.vfyjxf.nee.network.packet.PacketStackCountChange;
 import com.github.vfyjxf.nee.utils.GuiUtils;
+import com.github.vfyjxf.nee.utils.ItemUtils;
 
+import appeng.api.events.GuiScrollEvent;
+import appeng.client.gui.AEBaseGui;
+import appeng.client.gui.implementations.GuiCraftConfirm;
+import appeng.client.gui.implementations.GuiInterface;
+import appeng.client.gui.implementations.GuiPatternTerm;
+import appeng.container.implementations.ContainerCraftConfirm;
+import appeng.container.implementations.ContainerPatternTerm;
+import appeng.container.slot.SlotFake;
+import codechicken.nei.NEIClientConfig;
+import codechicken.nei.PositionedStack;
+import codechicken.nei.VisiblityData;
+import codechicken.nei.api.INEIGuiHandler;
+import codechicken.nei.api.TaggedInventoryArea;
 import cpw.mods.fml.common.Loader;
 import cpw.mods.fml.common.ObfuscationReflectionHelper;
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
@@ -231,5 +241,88 @@ public class GuiEventHandler implements INEIGuiHandler {
     @Override
     public boolean hideItemPanelSlot(GuiContainer gui, int x, int y, int w, int h) {
         return false;
+    }
+
+    /**
+     * Prevent the scroll bar from being triggered when modifying the number of items This method is not intended to be
+     * called by NEE. Do not use this method for any reason.
+     */
+    @SubscribeEvent
+    public boolean handleMouseWheelInput(GuiScrollEvent event) {
+        Minecraft mc = Minecraft.getMinecraft();
+        boolean isPatternTerm = isPatternTerm(mc.currentScreen);
+        boolean isInterface = event.guiScreen instanceof GuiInterface;
+        if (isPatternTerm || isInterface) {
+            Slot currentSlot = (event.guiScreen).getSlotAtPosition(event.mouseX, event.mouseY);
+            if (currentSlot instanceof SlotFake && currentSlot.getHasStack()) {
+                // try to change current itemstack to next ingredient;
+                if (Keyboard.isKeyDown(NEIClientConfig.getKeyBinding("nee.ingredient"))
+                        && GuiUtils.isCraftingSlot(currentSlot)) {
+                    handleRecipeIngredientChange(event.guiScreen, currentSlot, event.scrollAmount);
+                    return true;
+                } else if (Keyboard.isKeyDown(NEIClientConfig.getKeyBinding("nee.count"))) {
+                    NEENetworkHandler.getInstance()
+                            .sendToServer(new PacketStackCountChange(currentSlot.slotNumber, event.scrollAmount));
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+
+    private static void handleRecipeIngredientChange(GuiContainer gui, Slot currentSlot, int dWheel) {
+        List<Integer> craftingSlots = new ArrayList<>();
+        int currentSlotIndex = currentSlot.getSlotIndex();
+        PositionedStack currentIngredients = NEECraftingHandler.ingredients.get(INPUT_KEY + currentSlotIndex);
+        if (currentIngredients != null && currentIngredients.items.length > 1) {
+            int currentStackIndex = ItemUtils.getIngredientIndex(currentSlot.getStack(), currentIngredients);
+            if (currentStackIndex >= 0) {
+                for (int j = 0; j < Math.abs(dWheel); j++) {
+                    currentStackIndex += (dWheel < 0) ? -1 : 1;
+                    if (currentStackIndex >= currentIngredients.items.length) {
+                        currentStackIndex = 0;
+                    } else if (currentStackIndex < 0) {
+                        currentStackIndex = currentIngredients.items.length - 1;
+                    }
+
+                    ItemStack currentStack = currentIngredients.items[currentStackIndex].copy();
+                    currentStack.stackSize = currentSlot.getStack().stackSize;
+
+                    if (NEEConfig.allowSynchronousSwitchIngredient) {
+                        for (Slot slot : getCraftingSlots(gui)) {
+
+                            PositionedStack slotIngredients = NEECraftingHandler.ingredients
+                                    .get(INPUT_KEY + slot.getSlotIndex());
+
+                            boolean areItemStackEqual = currentSlot.getHasStack() && slot.getHasStack()
+                                    && currentSlot.getStack().isItemEqual(slot.getStack())
+                                    && ItemStack.areItemStackTagsEqual(currentSlot.getStack(), slot.getStack());
+
+                            boolean areIngredientEqual = slotIngredients != null
+                                    && currentIngredients.contains(slotIngredients.items[0]);
+
+                            if (areItemStackEqual && areIngredientEqual) {
+                                craftingSlots.add(slot.slotNumber);
+                            }
+                        }
+                    } else {
+                        craftingSlots.add(currentSlot.slotNumber);
+                    }
+                    NEENetworkHandler.getInstance()
+                            .sendToServer(new PacketSlotStackChange(currentStack, craftingSlots));
+                }
+            }
+        }
+    }
+
+    @SuppressWarnings("unchecked")
+    private static List<Slot> getCraftingSlots(GuiContainer gui) {
+        List<Slot> craftingSlots = new ArrayList<>();
+        for (Slot slot : (List<Slot>) gui.inventorySlots.inventorySlots) {
+            if (GuiUtils.isCraftingSlot(slot)) {
+                craftingSlots.add(slot);
+            }
+        }
+        return craftingSlots;
     }
 }
